Update README.md
# V2V-communication-
It is the project about communication between one vehicle to another vehicle, where in milatary Area some restriction would be there where in no object(man, vehicle, animal) Should not be entered in that Area.
 To detect any object is entered in that perticular Area that should be sensed by field sensors (temperature sensor, piR sensor, hall sensor)That information is sent to base stations through wireless communication
 using gateway and router, then the base stations sends the GSM commands to vehicles to move and ultrasonic sensor and camera module Fixed on the vehicle, whenever the object dectected between the path 
camera module capture that object and send to base station and other vehicles or soldiers. Then vehicle goes to the that restricted Area takes the picture of that Area and send to the base station.